Refactors project_utils module into core.project
