+++ docs/source/vara-ts-api/projects.rst
@@ -31,7 +31,7 @@ template for your own project::
 
     from varats.paper_mgmt.paper_config import project_filter_generator
     from varats.utilss.settings import bb_cfg
-    from varats.utilss.project_util import (
+    from varats.project.project_util import (
         ProjectBinaryWrapper,
         wrap_paths_to_binaries,
     )
