+++ varats-core/setup.py
@@ -21,6 +21,7 @@ setup(
         "requests>=2.23.0",
         "packaging>=20.1",
         "requests_cache>=0.5.2",
+        "pygit2>=0.28.2",
     ],
     author="Florian Sattler",
     author_email="sattlerf@cs.uni-saarland.de",
