+++ varats/varats/plots/blame_interaction_degree.py
@@ -18,7 +18,7 @@ from varats.plots.bug_annotation import draw_bugs
 from varats.plots.cve_annotation import draw_cves
 from varats.plots.plot import Plot, PlotDataEmpty
 from varats.plots.repository_churn import draw_code_churn_for_revisions
-from varats.utilss.project_util import get_project_cls_by_name
+from varats.project.project_util import get_project_cls_by_name
 
 LOG = logging.getLogger(__name__)
