+++ varats/varats/projects/c_projects/gravity.py
@@ -13,14 +13,14 @@ from benchbuild.utils.settings import get_number_of_jobs
 from plumbum import local
 
 from varats.paper_mgmt.paper_config import project_filter_generator
-from varats.provider.cve.cve_provider import CVEProviderHook
-from varats.utilss.project_util import (
+from varats.project.project_util import (
     ProjectBinaryWrapper,
     get_all_revisions_between,
     wrap_paths_to_binaries,
     get_local_project_git_path,
     BinaryType,
 )
+from varats.provider.cve.cve_provider import CVEProviderHook
 from varats.utilss.settings import bb_cfg
