+++ varats/varats/projects/cpp_projects/doxygen.py
@@ -9,7 +9,7 @@ from plumbum import local
 from plumbum.path.utils import delete
 
 from varats.paper_mgmt.paper_config import project_filter_generator
-from varats.utilss.project_util import (
+from varats.project.project_util import (
     ProjectBinaryWrapper,
     wrap_paths_to_binaries,
     BinaryType,
