+++ varats/varats/projects/test_projects/example_test_repo.py
@@ -5,7 +5,7 @@ import benchbuild as bb
 from plumbum import local
 
 from varats.paper_mgmt.paper_config import project_filter_generator
-from varats.utilss.project_util import (
+from varats.project.project_util import (
     VaraTestRepoSource,
     ProjectBinaryWrapper,
     wrap_paths_to_binaries,
