+++ varats/varats/tools/driver_casestudy.py
@@ -21,11 +21,11 @@ from varats.paper_mgmt.case_study import (
     generate_case_study,
 )
 from varats.paper_mgmt.paper_config import get_paper_config
+from varats.project.project_util import get_local_project_git_path
 from varats.provider.release.release_provider import ReleaseType
 from varats.report.report import FileStatusExtension, MetaReport
 from varats.tools.commit_map import create_lazy_commit_map_loader
 from varats.utilss.cli_util import cli_list_choice, initialize_cli_tool
-from varats.utilss.project_util import get_local_project_git_path
 from varats.utilss.settings import vara_cfg
 
 LOG = logging.getLogger(__name__)
