+++ varats/varats/utilss/experiment_util.py
@@ -15,8 +15,8 @@ from benchbuild.utils.cmd import prlimit
 from plumbum.commands import ProcessExecutionError
 
 from varats.data.revisions import get_tagged_revisions
+from varats.project.project_util import ProjectBinaryWrapper
 from varats.report.report import BaseReport, FileStatusExtension
-from varats.utilss.project_util import ProjectBinaryWrapper
 from varats.utilss.settings import vara_cfg, bb_cfg
